slicing layout weather. belum sempurna
