+++ app/src/main/res/drawable/border_weather.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <stroke android:color="#FFFFFF" android:width="2dp"/>
+    <corners android:radius="10dp" />
+</shape>
