+++ app/src/main/res/layout/activity_weather.xml
@@ -4,16 +4,166 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="#F6F6F6"
     tools:context=".ui.home.WeatherActivity">
 
+    <ImageView
+        android:id="@+id/btnBackInWeather"
+        android:layout_width="24dp"
+        android:layout_height="24dp"
+        android:layout_marginLeft="16dp"
+        android:layout_marginTop="16dp"
+        android:src="@drawable/baseline_arrow_back_24"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
     <TextView
+        android:id="@+id/textLocation"
+        android:layout_width="57dp"
+        android:layout_height="22dp"
+        android:layout_marginLeft="160dp"
+        android:layout_marginTop="16dp"
+        android:text="Jakarta"
+        android:textSize="16sp"
+        app:layout_constraintLeft_toLeftOf="@id/btnBackInWeather"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="HEllO"
-        android:textSize="55sp"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
+        android:layout_marginTop="12dp"
+        android:layout_marginRight="24dp"
+        android:src="@drawable/baseline_add_24"
         app:layout_constraintRight_toRightOf="parent"
-        tools:layout_editor_absoluteY="264dp" />
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <FrameLayout
+        android:id="@+id/frameKondisiCuaca"
+        android:layout_width="328dp"
+        android:layout_height="138dp"
+        android:layout_marginLeft="40dp"
+        android:layout_marginTop="60dp"
+        android:background="@drawable/border_weather"
+        android:backgroundTint="#FFFFFF"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/textLocation">
+
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:id="@+id/relativeSuhu">
+
+            <TextView
+                android:id="@+id/textSuhu"
+                android:layout_width="113sp"
+                android:layout_height="81sp"
+                android:text="35°"
+                android:textSize="64sp"
+                android:layout_marginLeft="12dp"
+                android:layout_marginTop="10dp"
+                android:textColor="@color/PrimaryBrand"
+                android:textAlignment="center"/>
+
+            <androidx.constraintlayout.utils.widget.ImageFilterView
+                android:id="@+id/iconSuhu"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:src="@drawable/sun"
+                android:layout_marginTop="8dp"
+                android:layout_below="@+id/textSuhu"
+                android:layout_marginLeft="22dp"/>
+
+            <TextView
+                android:id="@+id/kondisiSuhu"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Cerah"
+                android:textSize="16sp"
+                android:textStyle="bold"
+                android:layout_marginTop="10dp"
+                android:textAlignment="center"
+                android:gravity="center"
+                android:layout_below="@id/textSuhu"
+                android:layout_toRightOf="@id/iconSuhu"
+                android:layout_marginLeft="4dp"/>
+
+            <TextView
+                android:id="@+id/ringkasanHarian"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_alignParentRight="true"
+                android:layout_marginLeft="66dp"
+                android:layout_marginTop="18dp"
+                android:layout_marginRight="52dp"
+                android:text="Ringkasan Harian"
+                android:textSize="14sp" />
+
+            <TextView
+                android:layout_width="167dp"
+                android:layout_height="wrap_content"
+                android:text="Hari ini cerah, Yuk berikan pupuk dan vitamin serta pantau kelembaban tanaman Anda!"
+                android:layout_alignParentRight="true"
+                android:layout_below="@+id/ringkasanHarian"
+                android:textAlignment="center"
+                android:layout_marginTop="8dp"
+                android:layout_marginRight="24dp"/>
+        </RelativeLayout>
+    </FrameLayout>
+
+
+    <FrameLayout
+        android:id="@+id/frameKelembapan"
+        android:layout_width="150dp"
+        android:layout_height="150dp"
+        android:layout_marginLeft="16dp"
+        android:background="@drawable/border_weather"
+        android:backgroundTint="#FFFFFF"
+        android:layout_marginTop="120dp"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/frameKondisiCuaca">
+
+        <RelativeLayout
+            android:id="@+id/relativeKelembapan"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <TextView
+                android:id="@+id/textKelembapan"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_centerHorizontal="true"
+                android:layout_marginTop="8dp"
+                android:drawableStart="@drawable/embun"
+                android:drawablePadding="4dp"
+                android:text="Kelembapan"
+                android:textSize="14sp" />
+
+            <TextView
+                android:id="@+id/nilaiKelembapan"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/textKelembapan"
+                android:layout_centerHorizontal="true"
+                android:layout_marginTop="6dp"
+                android:text="85 %"
+                android:textColor="@color/PrimaryBrand"
+                android:textSize="40sp" />
+
+            <TextView
+                android:id="@+id/textTitikEmbun"
+                android:layout_width="95dp"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/nilaiKelembapan"
+                android:layout_centerHorizontal="true"
+                android:layout_marginTop="8dp"
+                android:text="Titik embun saat ini 17°C"
+                android:textSize="10sp" />
+
+
+        </RelativeLayout>
+
+
+    </FrameLayout>
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
