+++ app/src/main/res/layout/fragment_error404.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.Error404Fragment">
+    tools:context=".ui.home.Error404Fragment">
 
 
     <ImageView
